services/anomaly/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
paho-mqtt>=2.0
pyyaml>=6.0
numpy>=1.24
torch>=2.1.0
loguru>=0.7.0
